// src/shared/api/generated/openapi.d.ts
export interface paths {
// ...
  '/program/project/{id}': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    /** 프로젝트 수정 */
    put: operations['updateProject'];
    post?: never;
    /** 프로젝트 삭제 */
    delete: operations['deleteProject'];
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  '/admin/program/project/{id}': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    /** 프로젝트 수정 */
    put: operations['updateProject_1'];
    post?: never;
    /** 프로젝트 삭제 */
    delete: operations['deleteProject_1'];
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/admin/program/history/{id}': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    /** 연혁 수정 */
    put: operations['updateHistory'];
    post?: never;
    /** 연혁 삭제 */
    delete: operations['deleteHistory'];
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  '/program/upload/image': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /** 이미지 업로드 */
    post: operations['uploadImage'];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/program/project': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 프로젝트 목록 */
    get: operations['getProjects'];
    put?: never;
    /** 프로젝트 생성 */
    post: operations['createProject'];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  '/admin/program/history': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    get?: never;
    put?: never;
    /** 연혁 생성 */
    post: operations['getHistories'];
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
  '/user': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 유저 목록 */
    get: operations['getUsers_1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/recruit/latest': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 최근 모집 */
    get: operations['getLatestRecruit'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/program/study': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 모든 게시글 목록 */
    get: operations['getStudies'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/program/study/{category}': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 카테고리 게시글 목록 */
    get: operations['getStudies_1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/program/study/category': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 모든 카테고리 목록 */
    get: operations['getCategories'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/program/history': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 연혁 목록 */
    get: operations['getHistories_1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
  '/program/activity': {
    parameters: {
      query?: never;
      header?: never;
      path?: never;
      cookie?: never;
    };
    /** 활동 목록 */
    get: operations['getActivities_1'];
    put?: never;
    post?: never;
    delete?: never;
    options?: never;
    head?: never;
    patch?: never;
    trace?: never;
  };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
  schemas: {
// ...
    CreateProjectRequest: {
      title: string;
      image: string;
      description: string;
      link: string;
    };
// ...
    CreateHistoryRequest: {
      title: string;
      image: string;
      /** Format: date */
      date: string;
    };
// ...
    UsersResponse: {
      users?: components['schemas']['PublicUserResponse'][];
    };
// ...
  };
  responses: never;
  parameters: never;
  requestBodies: never;
  headers: never;
  pathItems: never;
}